requirements.txt
Scrapy==2.5.0
pymongo==3.12.0
pika==1.2.0
jsonschema==3.2.0
crochet==2.0.0
freezegun==1.1.0
python-dateutil==2.8.2
beautifulsoup4==4.10.0
elasticsearch==7.15.1
requests==2.26.0
html5lib==1.1
python-magic==0.4.24
oci==2.53.0
unidecode==1.3.2
capmonster_python==2.2.0
treq==22.2.0
ipdb==0.13.9
html2text==2020.1.16
anticaptchaofficial==1.0.47
spidermon[monitoring,validation]
log4mongo